fix: integral cast
Fix integral cast when src and dst have different sign.

+++ soul/src/core/type.h
@@ -310,8 +310,8 @@ namespace Soul {
 		SOUL_REQUIRE(is_integral_v<IntegralSrc>)
 	>
 	IntegralDst Cast(IntegralSrc src) {
-		SOUL_ASSERT(0, src <= std::numeric_limits<IntegralDst>::max(), "Source value is larger than the destintation type maximum!");
-		SOUL_ASSERT(0, src >= std::numeric_limits<IntegralDst>::min(), "Source value is smaller than the destination type minimum!");
+		SOUL_ASSERT(0, uint64(src) <= std::numeric_limits<IntegralDst>::max(), "Source value is larger than the destintation type maximum!");
+		SOUL_ASSERT(0, int64(src) >= std::numeric_limits<IntegralDst>::min(), "Source value is smaller than the destination type minimum!");
 		return IntegralDst(src);
 	}
